+++ packages/lex_gen/lib/src/commands/lex_command_generator.dart
@@ -9,14 +9,19 @@ import 'dart:io';
 import 'package:lexicon/lexicon.dart';
 
 // Project imports:
+import '../config.dart';
+import '../model/lex_def_kind.dart';
 import 'rule.dart';
 import 'types/lex_command.dart';
 import 'types/lex_parameter.dart';
 import 'types/lex_parent_command.dart';
 import 'types/lex_root_command.dart';
 
-void generateLexCommands(final List<LexiconDoc> docs) {
-  _cleanWorkspace();
+void generateLexCommands(
+  final LexCommandRuleConfig config,
+  final List<LexiconDoc> docs,
+) {
+  _cleanWorkspace(config);
 
   final result = <String, List<LexCommand>>{};
   for (final doc in docs) {
@@ -34,7 +39,7 @@ void generateLexCommands(final List<LexiconDoc> docs) {
               query.parameters?.requiredProperties,
               query.parameters?.properties,
             ),
-            isQuery: true,
+            kind: LexCommandKind.query,
           ),
         );
       } else if (def.value is ULexUserTypeXrpcProcedure) {
@@ -51,7 +56,7 @@ void generateLexCommands(final List<LexiconDoc> docs) {
               doc.id,
               procedure.description,
               const [],
-              isProcedure: true,
+              kind: LexCommandKind.procedure,
             ),
           );
 
@@ -68,7 +73,7 @@ void generateLexCommands(final List<LexiconDoc> docs) {
               procedure.description,
               const [],
               encoding: input.encoding,
-              isBlobProcedure: true,
+              kind: LexCommandKind.blobProcedure,
             ),
           );
 
@@ -86,7 +91,7 @@ void generateLexCommands(final List<LexiconDoc> docs) {
               doc.id,
               procedure.description,
               const [],
-              isProcedure: true,
+              kind: LexCommandKind.procedure,
               isRawJsonBody: true,
             ),
           );
@@ -101,7 +106,7 @@ void generateLexCommands(final List<LexiconDoc> docs) {
             doc.id,
             procedure.description,
             _getParameters(object.requiredProperties, object.properties),
-            isProcedure: true,
+            kind: LexCommandKind.procedure,
           ),
         );
       } else if (def.value is ULexUserTypeXrpcSubscription) {
@@ -117,7 +122,7 @@ void generateLexCommands(final List<LexiconDoc> docs) {
             record.description,
             _getParameters(object.requiredProperties, object.properties),
             rkey: record.key,
-            isRecord: true,
+            kind: LexCommandKind.record,
           ),
         );
       }
@@ -127,13 +132,18 @@ void generateLexCommands(final List<LexiconDoc> docs) {
   final parentCommands = <LexParentCommand>[];
   for (final entry in result.entries) {
     for (final command in entry.value) {
-      File(getAbsoluteFilePath(command.lexiconId.toString()))
+      File(getAbsoluteFilePath(config, command.lexiconId.toString()))
         ..createSync(recursive: true)
         ..writeAsStringSync(command.format());
     }
 
     final parentCommand = LexParentCommand(entry.key, entry.value);
-    File(getAbsoluteFilePathForParent(parentCommand.lexiconId.toString()))
+    File(
+        getAbsoluteFilePathForParent(
+          config,
+          parentCommand.lexiconId.toString(),
+        ),
+      )
       ..createSync(recursive: true)
       ..writeAsStringSync(parentCommand.format());
 
@@ -141,7 +151,7 @@ void generateLexCommands(final List<LexiconDoc> docs) {
   }
 
   final rootCommand = LexRootCommand(parentCommands);
-  File('${getHomeDir()}/lex_commands.dart')
+  File('${config.homeDir}/lex_commands.dart')
     ..createSync(recursive: true)
     ..writeAsStringSync(rootCommand.format());
 }
@@ -171,8 +181,8 @@ List<LexParameter> _getParameters(
   return parameters;
 }
 
-void _cleanWorkspace() {
-  final dir = Directory(getHomeDir());
+void _cleanWorkspace(final LexCommandRuleConfig config) {
+  final dir = Directory(config.homeDir);
   if (dir.existsSync()) dir.deleteSync(recursive: true);
 }